Fix: GoLang bindings (#70)
* golang: correct data type in function comments

Update the comments for CosineF32, InnerF32, and SqEuclideanF32
functions in the simsimd.go file. The comments incorrectly stated
that these functions compute the similarity between two i8 vectors.
They actually compute the similarity between two f32 vectors. This
change ensures that the comments accurately reflect the
functionality of the code.

* golang: change simsimd header file path

The path to the simsimd header file is updated. The header file is
in the 'include' directory.

* golang: link math library in golang simsimd

The math library is now linked in the golang simsimd file. The change was
necessary to resolve undefined references to `logf` and `sqrtf`.

* golang: update cosine_f32 function in simsimd.go

The cosine_f32 function in simsimd.go has been updated to use
simsimd_metric_punned instead of simsimd_neon_f32_cos. This change was
necessary to ensure compatibility across different SIMD architectures.
The previous implementation was specific to NEON, which could cause
issues on systems that do not support it.

This resolves the error:

```
./simsimd.go: In function 'cosine_f32':
./simsimd.go:10:115: warning: implicit declaration of function 'simsimd_neon_f32_cos'; did you mean 'simsimd_serial_f32_cos'?  [-Wimplicit-function-declaration]
```

* golang: add length check to vector operations in simsimd.go

Ensure that input vectors to the CosineI8, CosineF32, InnerI8, InnerF32,
SqEuclideanI8, and SqEuclideanF32 functions have the same length. This
prevents potential runtime errors when the functions are called with
vectors of different lengths. The functions now panic with a descriptive
message if the lengths do not match.