Improve: `Uint8Array` support in JS
NodeJS type system doesn't define `float16`,
so we don't currently have a way of detecting
such arrays.

In Python bindings we allow passing an additional
string parameter together with `uint16` arrays to
override the logical type for the missing `bfloat16`.
In JS, for now, it's wiser to wait for the NAPI to
define the missing `napi_typedarray_type` value.

https://github.com/nodejs/node/issues/55574#issuecomment-2442262127
https://github.com/nodejs/node/issues/52416#issuecomment-2043448555