Add: Same-type element-wise ops
New classes are added to the Python SDK:
public NDArray and NDIndex and internal
BufferOrScalarArgument. Those can be used
for high-dimensional tensor processing with
up to 64 dimensions, as opposed to 32 in
NumPy.

The new interface handles mixed-precision
much better, allowing to override the type
spec of every tensor individually. That feature
is in preview until the actual implementation
in subsequent commits.

No LibC is needed anymore for rounding floats
to 8-bit integers in elementwise ops. New 16-,
32-, and 64-bit integer element-wise kernels
are added for compatibility with NumPy. Serial
for now.

SimSIMD implementation is supposed to be
more efficient when at least a few continuous
dimensions are present.