fix: Fix bug in SIMD accelerated algorithm
The SIMD version was reporting the number of matched items between the Sparse vectors
incorrectly. Fix the bug and remove unnecessary variables from the functions

Add a bin file with custom loop for calculating benchmarks. This is faster to run
and does averaging of timing runs for various vector sizes